Feat: Utilize initWorkspaceImages and getImageInfo function
